Additional Code Examples
RunQuery is a simple script to automatically pass a query to my selected database, rather than writing the connection code each time.
BuildForm is a script that assembles all the data from the (proprietary) database and puts it together into a simple HTML file. This is a versatile function since, with some expansion, could be used to build every aspect of a web page from the data driven back end of the MySQL database.